feat: Add approve button to targets list
